Refactor E2E test code into separate files
